data-transformation/requirements.txt
tabula-py
pandas
JPype1